a bit of organization, added Future to App
