+++ framework/src/play-functional/src/main/scala/play/api/libs/functional/Functors.scala
@@ -10,6 +10,22 @@ trait Functor[M[_]] extends Variant[M]{
 
 }
 
+object Functor {
+
+  implicit def functorOption: Functor[Option] = new Functor[Option] {
+
+    def fmap[A,B](a:Option[A], f: A => B):Option[B] = a.map(f)
+
+  }
+
+  implicit def functorFuture(implicit ec: scala.concurrent.ExecutionContext): Functor[scala.concurrent.Future] = new Functor[scala.concurrent.Future] {
+
+    def fmap[A,B](a:scala.concurrent.Future[A], f: A => B):scala.concurrent.Future[B] = a.map(f)
+
+  }
+
+}
+
 trait InvariantFunctor[M[_]] extends Variant[M]{
 
   def inmap[A,B](m: M[A], f1: A => B, f2: B => A): M[B]
