+++ framework/src/play-functional/src/main/scala/play/api/libs/functional/Products.scala
@@ -10,6 +10,16 @@ trait FunctionalCanBuild[M[_]]{
 
 }
 
+object FunctionalCanBuild {
+
+  implicit def fuanctionalCanBuildApplicative[M[_]](implicit app: Applicative[M]): FunctionalCanBuild[M] = new FunctionalCanBuild[M] {
+
+    def apply[A,B](a: M[A], b: M[B]): M[A~B] = app.apply(app.map[A, B => A ~ B](a, a => ((b: B) => new ~(a,b))),b)
+
+  }
+
+}
+
 class FunctionalBuilderOps[M[_],A](ma: M[A])(implicit fcb: FunctionalCanBuild[M]){
 
   def ~[B](mb: M[B]): FunctionalBuilder[M]#CanBuild2[A,B] = {
