+++ framework/src/play-functional/src/main/scala/play/api/libs/functional/syntax/package.scala
@@ -17,28 +17,6 @@ object `package` {
 
   implicit def toFunctionalBuilderOps[M[_],A](a: M[A])(implicit fcb: FunctionalCanBuild[M]) = new FunctionalBuilderOps[M,A](a)(fcb)
 
-  implicit def functionalCanBuildApplicative[M[_]](implicit app: Applicative[M]): FunctionalCanBuild[M] = new FunctionalCanBuild[M] {
-
-    def apply[A,B](a: M[A], b: M[B]): M[A~B] = app.apply(app.map[A, B => A ~ B](a, a => ((b: B) => new ~(a,b))),b)
-
-  }
-
-  implicit def functorOption: Functor[Option] = new Functor[Option] {
-
-    def fmap[A,B](a:Option[A], f: A => B):Option[B] = a.map(f)
-
-  }
-
-  implicit def applicativeOption: Applicative[Option] = new Applicative[Option]{
-
-    def pure[A](a: A):Option[A] = Some(a)
-
-    def map[A,B](m:Option[A], f: A => B):Option[B] = m.map(f)
-
-    def apply[A,B](mf:Option[A => B], ma: Option[A]):Option[B] = mf.flatMap(f => ma.map(f))
-
-  }
-
   implicit def toFunctorOps[M[_], A](ma: M[A])(implicit fu: Functor[M]): FunctorOps[M, A] = new FunctorOps(ma)
   implicit def toContraFunctorOps[M[_], A](ma: M[A])(implicit fu: ContravariantFunctor[M]): ContravariantFunctorOps[M, A] = new ContravariantFunctorOps(ma)
   implicit def toInvariantFunctorOps[M[_], A](ma: M[A])(implicit fu: InvariantFunctor[M]): InvariantFunctorOps[M, A] = new InvariantFunctorOps(ma)
